backend/pyproject.toml
[tool.poetry]
name = "infilmation"
version = "0.1.0"
description = ""
authors = ["Dom Batten <dominic.batten@googlemail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
ormar = "^0.10.22"
psycopg2-binary = "^2.9.1"
asyncpg = "^0.24.0"
alembic = "^1.7.4"
phylm = "^4.0.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
coverage = {extras = ["toml"], version = "^6.1"}
safety = "^1.10.3"
mypy = "^0.910"
typeguard = "^2.12.1"
xdoctest = {extras = ["colors"], version = "^0.15.5"}
pre-commit = "^2.13.0"
flake8 = "^4.0.1"
black = "^21.10b0"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^21.4.3"
flake8-docstrings = "^1.6.0"
flake8-rst-docstrings = "^0.2.3"
pep8-naming = "^0.12.0"
darglint = "^1.8.0"
reorder-python-imports = "^2.6.0"
pre-commit-hooks = "^4.0.1"
Pygments = "^2.9.0"
types-toml = "^0.10.1"
types-requests = "^2.25.11"
pylint = "^2.11.1"
pytest-dotenv = "^0.5.2"
sqlalchemy-stubs = "^0.4"
pytest-asyncio = "^0.16.0"

[tool.mypy]
strict = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
plugins = "sqlmypy"

[tool.isort]
profile = "open_stack"

[tool.pytest.ini_options]
env_override_existing_values = true
env_files = [".env",".env.test"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
